social-media-agent/requirements.txt
tweepy>=4.14,<5.0
requests>=2.28,<3.0
requests-oauthlib>=1.3,<2.0
python-dateutil>=2.8,<3.0
